refactor: fix last remarks
